Made Portals reusable, modularized packages
Found out that the bugs relating to "Stdin/Stdout already set" had to
do with the fact that exec.Cmd is actually meant for one-time use, and
cannot be restarted once it has been stopped.

In response to this, I gave Portal the ability to reload its internal
exec.Cmd from a reference struct; it now does this at the start of
every action.

Also, removed Travis CI tests for Linux platforms (for now)—that X
server seems like a real hastle to set up.